GITA MASALA

+++ src/swichcase/switch5.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class switch5 {
     public static void main(String[] args) {
         Scanner scanner=new Scanner(System.in);
-        int a,b;
+        double a,b;
         System.out.println("1 ->  qo`shish");
         System.out.println("2 ->  ayrish");
         System.out.println("3 ->  bo`lish");
@@ -15,36 +15,36 @@ public class switch5 {
         switch (choose){
             case 1->{
                 System.out.println("1- Sonni kiriting:");
-                a= scanner.nextInt();
+                a= scanner.nextDouble();
                 System.out.println("2- Sonni kiriting:");
-                b=scanner.nextInt();
-                int natija=a+b;
+                b=scanner.nextDouble();
+                int natija= (int) (a+b);
                 System.out.printf("%d + %d = %d",a,b,natija);
                 break;
             }
             case 2->{
                 System.out.println("1- Sonni kiriting:");
-                a= scanner.nextInt();
+                a= scanner.nextDouble();
                 System.out.println("2- Sonni kiriting:");
-                b=scanner.nextInt();
-                int natija=a-b;
+                b=scanner.nextDouble();
+                int natija= (int) (a-b);
                 System.out.printf("%d - %d = %d",a,b,natija);
                 break;
             }
             case 3->{
                 System.out.println("1- Sonni kiriting:");
-                a= scanner.nextInt();
+                a= scanner.nextDouble();
                 System.out.println("2- Sonni kiriting:");
-                b=scanner.nextInt();
-                int natija=a/b;
+                b=scanner.nextDouble();
+                int natija= (int) (a/b);
                 System.out.printf("%d / %d = %d",a,b,natija);
             }
             case 4->{
                 System.out.println("1- Sonni kiriting:");
-                a= scanner.nextInt();
+                a= scanner.nextDouble();
                 System.out.println("2- Sonni kiriting:");
-                b=scanner.nextInt();
-                int natija=a*b;
+                b=scanner.nextDouble();
+                int natija= (int) (a*b);
                 System.out.printf("%d * %d = %d",a,b,natija);
             }
             default -> {
